minor adjustments and thread locking trials.. still crashing rarely, due to some strange error, or locking issue as explained below:
Looks like the crash is related to this  errormessage "QSortFilterProxyModel: index from wrong model passed to mapFromSource"
Confirming this is that it looks like it crashes ONLY when connected to the broker AND thetableis sorted by clicking on a column header (TV Prof)
As a workaround I put the "TV-Chg %" column in front as QTableView sorts automatically by the first column when constructed.
So I have the correct  sorting but  still cannot sort during operation without being worried that the app anytime can crash - which is still annoying of course

But could not find any useful hints about it and the errormessage above.
Backtracing the resulting segmentationfault seems to point to these three lines in updateModel:

"       self.layoutAboutToBeChanged.emit()
        self.dataChanged.emit(self.createIndex(0, 0), self.createIndex(self.rowCount(0), self.columnCount(0)))
        self.layoutChanged.emit()
"

Currently no time  for further investigation

Maybe it is Qt, PyQt, dont know. Some difficult task for the future