Update README.md
at last: stablized thank God ;-) 
no more crashes occurred, probably because of the removal of the layoutAboutToBeChanged signals and something like that, 
only one signal will be sent (datachanged) if data is updated. So no more crashes happened since then.

Not the LAYOUT is changed if just some data changed. Keep this in mind, to be learned...